Fix tests

+++ 2019-adventofcode.com/aoc2019/aoc/lcmgcd_test.go
@@ -7,6 +7,6 @@ import (
 )
 
 func TestLcm(t *testing.T) {
-	assert.Equal(t, Lcm(6, 10), 30)
-	assert.Equal(t, Lcm(10, 6), 30)
+	assert.Equal(t, Lcm(6, 10), int64(30))
+	assert.Equal(t, Lcm(10, 6), int64(30))
 }
